Remove repeated calls for account name, instead use cached value in Config
